+++ app/hospital/[hos_id]/patients/page.tsx
@@ -1,13 +1,11 @@
 import SearchPatientContainer from '@/components/hospital/icu/header/register-dialog/search-patient/search-patient-containter'
 import PatientRegisterDialog from '@/components/hospital/patients/patient-register-dialog'
 import UploadPatientArea from '@/components/hospital/patients/upload-patient-area'
-import { getHosPatientCount } from '@/lib/services/patient/patient'
 
 export default async function HospitalPatientsPage(props: {
   params: Promise<{ hos_id: string }>
 }) {
   const params = await props.params
-  const totalPatientCount = await getHosPatientCount(params.hos_id)
 
   return (
     <div className="overflow-x-scroll p-2">
@@ -18,9 +16,9 @@ export default async function HospitalPatientsPage(props: {
       </div>
 
       <SearchPatientContainer
-        totalPatientCount={totalPatientCount}
         itemsPerPage={10}
         isIcu={false}
+        hosId={params.hos_id}
       />
     </div>
   )
