+++ components/hospital/icu/header/register-dialog/search-patient/search-patient-containter.tsx
@@ -42,7 +42,7 @@ export default function SearchPatientContainer({
 
   useEffect(() => {
     getHosPatientCount(hosId).then(setTotalPatientCount)
-  }, [])
+  }, [hosId])
 
   const handleInputChange = (e: React.ChangeEvent<HTMLInputElement>) => {
     setInputValue(e.target.value)
